Do not block orders on failed Avalara address validations
We don't want to block the checkout when the Avatax service is down,
or timeouts, or anything unexpected and unrelated to address
validation that could prevent the customer from checking out their
carts.

This is a quick and dirty solution to get us going, in case of errors
the class instance methods may not be fully coherent and working but
this is enough to keep moving the order forward without blocking the
checkout.

A better and more comprehensive solution would be to rework the way
errors are managed starting from `TaxSvc#validate_address` [1] but
this would require more extensive rework.

Original specs are retained but reworked so they can be integrated
in the next commit separately for clarity with new ones that verify
this change.

[1] https://github.com/watg/solidus_avatax_certified/blob/master/app/models/tax_svc.rb#L29-L36